refactor: update EnergyConsumptionChart
Pulls props into component as these will not change. Implements Chakra.
